+++ CHANGELOG.md
@@ -3,6 +3,13 @@ the <a href="https://github.com/marcglasberg/SameAppDifferentTech/blob/main/Mobi
 Async Redux App Example GitHub Repo</a> for a full-fledged example app showcasing the fundamentals
 and best practices._
 
+# 23.3.0-dev.1
+
+* `LocalPersist` and `LocalJsonPersist` now allow you to define the base directory by setting
+  the `findBaseDirectory` static field. The default is, as before, the application's documents
+  directory. Other options are the cache directory, the downloads directory, or another custom
+  directory.
+
 # 23.2.0
 
 * You can now use the `UnlimitedRetryCheckInternet` to check if there is internet when you run some